Added routes for spotify related API calls. Relocated endpoints from index.js to spotifyRoutes.js.
